# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Z Route Master API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface MessageResponse {
  message: string;
}

export interface LoginBody {
  username: string;
  password: string;
}

export interface AuthUser {
  username: string;
  displayName: string;
  plannerName: string;
  role: string;
}

export interface UserResponse {
  id: number;
  username: string;
  displayName: string;
  plannerName: string;
  role: string;
}

export interface CreateUserBody {
  username: string;
  password: string;
  displayName: string;
  plannerName: string;
}

export interface SiteData {
  id: string;
  name: string;
  code?: string;
  lat: number;
  lng: number;
  gov?: string;
  dist?: string;
  subdist?: string;
  key?: string;
  pwrclass?: string;
  vendor?: string;
  cat?: string;
  owner?: string;
  tl?: string;
  mw?: string;
  fiber?: string;
  dwdm?: string;
  olt?: string;
  ipran?: string;
  fttm?: string;
  epa?: string;
  freecool?: string;
  gen?: string;
  sharing?: string;
  ibs?: string;
}

export interface DayGroup {
  day: number;
  date: string;
  siteIds: string[];
}

export interface PlanData {
  id: number;
  teamName: string;
  plannerName: string;
  planName: string;
  color: string;
  km?: number;
  isNewSites?: boolean;
  hqSiteId?: string | null;
  dayGroups: DayGroup[];
  siteIds: string[];
}