# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Z Route Master API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: auth
    description: Authentication
  - name: users
    description: User management
  - name: sites
    description: Site data
  - name: plans
    description: Routing plans
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /auth/logout:
    post:
      operationId: logout
      tags: [auth]
      summary: Logout
      responses:
        "200":
          description: Logged out
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current session user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /users:
    get:
      operationId: getUsers
      tags: [users]
      summary: Get all users (admin only)
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserResponse"
    post:
      operationId: createUser
      tags: [users]
      summary: Create a new user (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserBody"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /users/{id}:
    delete:
      operationId: deleteUser
      tags: [users]
      summary: Delete a user (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  /sites:
    get:
      operationId: getSites
      tags: [sites]
      summary: Get all sites
      responses:
        "200":
          description: List of sites
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SiteData"
    delete:
      operationId: clearSites
      tags: [sites]
      summary: Delete all sites
      responses:
        "200":
          description: Cleared
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  /sites/bulk:
    post:
      operationId: bulkSaveSites
      tags: [sites]
      summary: Replace all sites with a new set
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: "#/components/schemas/SiteData"
      responses:
        "200":
          description: Saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  /plans:
    get:
      operationId: getPlans
      tags: [plans]
      summary: Get plans
      responses:
        "200":
          description: List of plans
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PlanData"
    delete:
      operationId: clearPlans
      tags: [plans]
      summary: Delete all plans
      responses:
        "200":
          description: Cleared
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  /plans/bulk:
    post:
      operationId: bulkSavePlans
      tags: [plans]
      summary: Replace all plans with a new set
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: "#/components/schemas/PlanData"
      responses:
        "200":
          description: Saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  /plans/{id}:
    put:
      operationId: updatePlan
      tags: [plans]
      summary: Update a plan
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlanData"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
    delete:
      operationId: deletePlan
      tags: [plans]
      summary: Delete a plan
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required: [status]

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]

    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]

    LoginBody:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
      required: [username, password]

    AuthUser:
      type: object
      properties:
        username:
          type: string
        displayName:
          type: string
        plannerName:
          type: string
        role:
          type: string
      required: [username, displayName, plannerName, role]

    UserResponse:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        displayName:
          type: string
        plannerName:
          type: string
        role:
          type: string
      required: [id, username, displayName, plannerName, role]

    CreateUserBody:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
        displayName:
          type: string
        plannerName:
          type: string
      required: [username, password, displayName, plannerName]

    SiteData:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        code:
          type: string
        lat:
          type: number
        lng:
          type: number
        gov:
          type: string
        dist:
          type: string
        subdist:
          type: string
        key:
          type: string
        pwrclass:
          type: string
        vendor:
          type: string
        cat:
          type: string
        owner:
          type: string
        tl:
          type: string
        mw:
          type: string
        fiber:
          type: string
        dwdm:
          type: string
        olt:
          type: string
        ipran:
          type: string
        fttm:
          type: string
        epa:
          type: string
        freecool:
          type: string
        gen:
          type: string
        sharing:
          type: string
        ibs:
          type: string
      required: [id, name, lat, lng]

    DayGroup:
      type: object
      properties:
        day:
          type: integer
        date:
          type: string
        siteIds:
          type: array
          items:
            type: string
      required: [day, date, siteIds]

    PlanData:
      type: object
      properties:
        id:
          type: integer
        teamName:
          type: string
        plannerName:
          type: string
        planName:
          type: string
        color:
          type: string
        km:
          type: number
        isNewSites:
          type: boolean
        hqSiteId:
          type: string
          nullable: true
        dayGroups:
          type: array
          items:
            $ref: "#/components/schemas/DayGroup"
        siteIds:
          type: array
          items:
            type: string
      required: [id, teamName, plannerName, planName, color, dayGroups, siteIds]