# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Z Route Master API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary Login
 */
export const LoginBody = zod.object({
  username: zod.string(),
  password: zod.string(),
});

export const LoginResponse = zod.object({
  username: zod.string(),
  displayName: zod.string(),
  plannerName: zod.string(),
  role: zod.string(),
});

/**
 * @summary Logout
 */
export const LogoutResponse = zod.object({
  message: zod.string(),
});

/**
 * @summary Get current session user
 */
export const GetMeResponse = zod.object({
  username: zod.string(),
  displayName: zod.string(),
  plannerName: zod.string(),
  role: zod.string(),
});

/**
 * @summary Get all users (admin only)
 */
export const GetUsersResponseItem = zod.object({
  id: zod.number(),
  username: zod.string(),
  displayName: zod.string(),
  plannerName: zod.string(),
  role: zod.string(),
});
export const GetUsersResponse = zod.array(GetUsersResponseItem);

/**
 * @summary Create a new user (admin only)
 */
export const CreateUserBody = zod.object({
  username: zod.string(),
  password: zod.string(),
  displayName: zod.string(),
  plannerName: zod.string(),
});

/**
 * @summary Delete a user (admin only)
 */
export const DeleteUserParams = zod.object({
  id: zod.coerce.number(),
});

export const DeleteUserResponse = zod.object({
  message: zod.string(),
});

/**
 * @summary Get all sites
 */
export const GetSitesResponseItem = zod.object({
  id: zod.string(),
  name: zod.string(),
  code: zod.string().optional(),
  lat: zod.number(),
  lng: zod.number(),
  gov: zod.string().optional(),
  dist: zod.string().optional(),
  subdist: zod.string().optional(),
  key: zod.string().optional(),
  pwrclass: zod.string().optional(),
  vendor: zod.string().optional(),
  cat: zod.string().optional(),
  owner: zod.string().optional(),
  tl: zod.string().optional(),
  mw: zod.string().optional(),
  fiber: zod.string().optional(),
  dwdm: zod.string().optional(),
  olt: zod.string().optional(),
  ipran: zod.string().optional(),
  fttm: zod.string().optional(),
  epa: zod.string().optional(),
  freecool: zod.string().optional(),
  gen: zod.string().optional(),
  sharing: zod.string().optional(),
  ibs: zod.string().optional(),
});
export const GetSitesResponse = zod.array(GetSitesResponseItem);

/**
 * @summary Delete all sites
 */
export const ClearSitesResponse = zod.object({
  message: zod.string(),
});

/**
 * @summary Replace all sites with a new set
 */
export const BulkSaveSitesBodyItem = zod.object({
  id: zod.string(),
  name: zod.string(),
  code: zod.string().optional(),
  lat: zod.number(),
  lng: zod.number(),
  gov: zod.string().optional(),
  dist: zod.string().optional(),
  subdist: zod.string().optional(),
  key: zod.string().optional(),
  pwrclass: zod.string().optional(),
  vendor: zod.string().optional(),
  cat: zod.string().optional(),
  owner: zod.string().optional(),
  tl: zod.string().optional(),
  mw: zod.string().optional(),
  fiber: zod.string().optional(),
  dwdm: zod.string().optional(),
  olt: zod.string().optional(),
  ipran: zod.string().optional(),
  fttm: zod.string().optional(),
  epa: zod.string().optional(),
  freecool: zod.string().optional(),
  gen: zod.string().optional(),
  sharing: zod.string().optional(),
  ibs: zod.string().optional(),
});
export const BulkSaveSitesBody = zod.array(BulkSaveSitesBodyItem);

export const BulkSaveSitesResponse = zod.object({
  message: zod.string(),
});

/**
 * @summary Get plans
 */
export const GetPlansResponseItem = zod.object({
  id: zod.number(),
  teamName: zod.string(),
  plannerName: zod.string(),
  planName: zod.string(),
  color: zod.string(),
  km: zod.number().optional(),
  isNewSites: zod.boolean().optional(),
  hqSiteId: zod.string().nullish(),
  dayGroups: zod.array(
    zod.object({
      day: zod.number(),
      date: zod.string(),
      siteIds: zod.array(zod.string()),
    }),
  ),
  siteIds: zod.array(zod.string()),
});
export const GetPlansResponse = zod.array(GetPlansResponseItem);

/**
 * @summary Delete all plans
 */
export const ClearPlansResponse = zod.object({
  message: zod.string(),
});

/**
 * @summary Replace all plans with a new set
 */
export const BulkSavePlansBodyItem = zod.object({
  id: zod.number(),
  teamName: zod.string(),
  plannerName: zod.string(),
  planName: zod.string(),
  color: zod.string(),
  km: zod.number().optional(),
  isNewSites: zod.boolean().optional(),
  hqSiteId: zod.string().nullish(),
  dayGroups: zod.array(
    zod.object({
      day: zod.number(),
      date: zod.string(),
      siteIds: zod.array(zod.string()),
    }),
  ),
  siteIds: zod.array(zod.string()),
});
export const BulkSavePlansBody = zod.array(BulkSavePlansBodyItem);

export const BulkSavePlansResponse = zod.object({
  message: zod.string(),
});

/**
 * @summary Update a plan
 */
export const UpdatePlanParams = zod.object({
  id: zod.coerce.number(),
});

export const UpdatePlanBody = zod.object({
  id: zod.number(),
  teamName: zod.string(),
  plannerName: zod.string(),
  planName: zod.string(),
  color: zod.string(),
  km: zod.number().optional(),
  isNewSites: zod.boolean().optional(),
  hqSiteId: zod.string().nullish(),
  dayGroups: zod.array(
    zod.object({
      day: zod.number(),
      date: zod.string(),
      siteIds: zod.array(zod.string()),
    }),
  ),
  siteIds: zod.array(zod.string()),
});

export const UpdatePlanResponse = zod.object({
  message: zod.string(),
});

/**
 * @summary Delete a plan
 */
export const DeletePlanParams = zod.object({
  id: zod.coerce.number(),
});

export const DeletePlanResponse = zod.object({
  message: zod.string(),
});