# lib/api-zod/src/generated/types/createUserBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Z Route Master API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateUserBody {
  username: string;
  password: string;
  displayName: string;
  plannerName: string;
}