# lib/api-zod/src/generated/types/dayGroup.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Z Route Master API
 * OpenAPI spec version: 0.1.0
 */

export interface DayGroup {
  day: number;
  date: string;
  siteIds: string[];
}