# lib/api-zod/src/generated/types/planData.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Z Route Master API
 * OpenAPI spec version: 0.1.0
 */
import type { DayGroup } from "./dayGroup";

export interface PlanData {
  id: number;
  teamName: string;
  plannerName: string;
  planName: string;
  color: string;
  km?: number;
  isNewSites?: boolean;
  hqSiteId?: string | null;
  dayGroups: DayGroup[];
  siteIds: string[];
}