# lib/api-zod/src/generated/types/siteData.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Z Route Master API
 * OpenAPI spec version: 0.1.0
 */

export interface SiteData {
  id: string;
  name: string;
  code?: string;
  lat: number;
  lng: number;
  gov?: string;
  dist?: string;
  subdist?: string;
  key?: string;
  pwrclass?: string;
  vendor?: string;
  cat?: string;
  owner?: string;
  tl?: string;
  mw?: string;
  fiber?: string;
  dwdm?: string;
  olt?: string;
  ipran?: string;
  fttm?: string;
  epa?: string;
  freecool?: string;
  gen?: string;
  sharing?: string;
  ibs?: string;
}