# lib/api-zod/src/generated/types/userResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Z Route Master API
 * OpenAPI spec version: 0.1.0
 */

export interface UserResponse {
  id: number;
  username: string;
  displayName: string;
  plannerName: string;
  role: string;
}